requirements.txt
kiteconnect>=4.2.0
python-dotenv>=1.0.0
pandas>=2.0.0
pymysql>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
apscheduler>=3.10.0
pytz>=2024.1
pyyaml>=6.0.0
